chore: move lock logic to buy conditions hook
